add MUI to navbar
